requirements.txt
appdirs==1.4.3
asn1crypto==1.1.0
attrdict==2.0.1
attrs==19.3.0
certifi==2019.9.11
cffi==1.13.0
chardet==3.0.4
Click==7.0
cmd2==0.9.4
colorama==0.4.1
cryptography==2.7
Flask==1.1.1
future==0.18.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
pycparser==2.19
pyperclip==1.7.0
python-telegram-bot==12.2.0
requests==2.22.0
schedule==0.6.0
six==1.12.0
tomcatmanager==0.14.0
tornado==6.0.3
urllib3==1.25.6
wcwidth==0.1.7
Werkzeug==0.16.0
